upgrade and move
